chore: apply arlrac/npm-package-template style with rollup + ava + nyc
