+++ package.json
@@ -5,19 +5,18 @@
   },
   "version": "0.0.0-semantic-release",
   "description": "kronos cluster node sample application",
-  "module": "src/cluster_node.js",
-  "main": "bin/cluster_node",
+  "module": "src/kronos-cluster-node.js",
+  "main": "bin/kronos-cluster-node",
   "scripts": {
-    "start": "node bin/cluster_node",
+    "start": "node bin/kronos-cluster-node",
     "cover": "nyc ava",
     "test": "ava",
-    "semantic-release": "semantic-release pre && npm publish && semantic-release post",
-    "docs": "jsdoc2md --configure doc/jsdoc.json -l off -t doc/README.hbs -f src/*.js >README.md",
     "pretest": "rollup -c tests/rollup.config.js",
-    "posttest": "npm run prepare && markdown-doctest",
     "precover": "rollup -c tests/rollup.config.js",
-    "prepare": "rollup -c -- src/cluster_node.js",
-    "prepublish": "undefined && chmod +x bin/kronos-cluster-node"
+    "posttest": "npm run prepare && markdown-doctest",
+    "docs": "jsdoc2md --configure doc/jsdoc.json -l off -t doc/README.hbs -f src/*.js >README.md",
+    "semantic-release": "semantic-release pre && npm publish && semantic-release post",
+    "prepare": "rollup -c -- src/kronos-cluster-node.js && chmod +x bin/kronos-cluster-node"
   },
   "engines": {
     "node": ">=8.1.4"
@@ -91,12 +90,13 @@
   ],
   "nyc": {
     "include": [
-      "dist/**/*.js"
+      "build/test-bundle.js",
+      "src/**/*.js"
     ],
     "reporter": [
       "lcov"
     ],
-    "report-dir": "./coverage"
+    "report-dir": "./build/coverage"
   },
   "template": {
     "repository": {
