+++ rollup.config.js
@@ -1,12 +1,15 @@
-/* jslint node: true, esnext: true */
-'use strict';
-
 import nodeResolve from 'rollup-plugin-node-resolve';
 import commonjs from 'rollup-plugin-commonjs';
+import pkg from './package.json';
 
 export default {
   banner: '#!/usr/bin/env node',
-  format: 'cjs',
+  targets: [
+    {
+      dest: pkg.bin['kronos-cluster-node'],
+      format: 'cjs'
+    }
+  ],
   plugins: [nodeResolve(), commonjs()],
   external: ['kronos-service-manager']
 };
