+++ src/util.js
@@ -1,7 +1,3 @@
-/* jslint node: true, esnext: true */
-
-'use strict';
-
 const fs = require('fs'),
   path = require('path'),
   glob = require('glob');
@@ -19,41 +15,52 @@ export function pglob(path, options) {
 }
 
 export function kronosModules() {
-  return pglob(path.join(__dirname, '..', 'node_modules/*/package.json'), {}).then(files => {
+  return pglob(
+    path.join(__dirname, '..', 'node_modules/*/package.json'),
+    {}
+  ).then(files => {
     const modules = [];
-    return Promise.all(files.map(file =>
-      new Promise((fullfill, reject) => {
-        fs.readFile(file, (err, data) => {
-          if (err) {
-            reject(`loading ${file}: ${err}`);
-            return;
-          }
-          try {
-            const p = JSON.parse(data);
-
-            if (p.keywords) {
-              if (p.keywords.find(k =>
-                  k === 'kronos-step' || k === 'kronos-service' || k === 'kronos-interceptor')) {
-                try {
-                  modules.push(require(p.name));
-                  fullfill();
-                  return;
-                } catch (e) {
-                  reject(`${file}: ${p.name} ${e}`);
-                }
+    return Promise.all(
+      files.map(
+        file =>
+          new Promise((fullfill, reject) => {
+            fs.readFile(file, (err, data) => {
+              if (err) {
+                reject(`loading ${file}: ${err}`);
+                return;
               }
-            }
-          } catch (e) {
-            reject(e);
-          }
+              try {
+                const p = JSON.parse(data);
 
-          fullfill();
-        });
-      })
-    )).then(results => modules);
+                if (p.keywords) {
+                  if (
+                    p.keywords.find(
+                      k =>
+                        k === 'kronos-step' ||
+                        k === 'kronos-service' ||
+                        k === 'kronos-interceptor'
+                    )
+                  ) {
+                    try {
+                      modules.push(require(p.name));
+                      fullfill();
+                      return;
+                    } catch (e) {
+                      reject(`${file}: ${p.name} ${e}`);
+                    }
+                  }
+                }
+              } catch (e) {
+                reject(e);
+              }
+
+              fullfill();
+            });
+          })
+      )
+    ).then(results => modules);
   });
 }
-
 
 export function assign(dest, value, attributePath) {
   const m = attributePath.match(/^(\w+)\.(.*)/);
